Server with token & react

+++ servernode/routes/answer.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Answer = require('../models/answer.model')
+let AuthToken = require('../models/authToken.model');
 
 ////GET REQUESTS
 
@@ -24,11 +25,17 @@ router.get('/findByRemark', (req,res,next) => {
 
 //http://localhost:5000/answers/findByUserId?id=5e500b859febd9351c7bdac2
 router.get('/findByUserId', (req,res,next) => {
-    Answer.find({ 
-        userId: req.query.id
+    AuthToken.findById(req.query.token )
+    .then((token) => {
+        Answer.find({ 
+            userId: token.userId
+        })
+        .then((answers) => res.status(200).json(answers))
+        .catch(err => res.status(400).json('Error:' + err))
+	})
+    .catch(err => {
+        res.status(401).json('Authentication Error: ' + err)
     })
-    .then((answers) => res.status(200).json(answers))
-    .catch(err => res.status(400).json('Error:' + err))
 });
 
 //http://localhost:5000/answers/sorted/date?order=1&skip=0&number=4
@@ -66,45 +73,67 @@ router.get('/sorted/pertinency', (req,res,next) => {
 
 //http://localhost:5000/answers/add
 router.route('/add').post((req, res) =>{
-    const userId = req.body.userId; 
-    const remarkId = req.body.remarkId;
-    const content = req.body.content;
-    const newAnswer = new Answer({ userId, remarkId, content })
-    newAnswer.save()
-    .then(() => res.status(200).json('Answer added.'))
-    .catch(err => res.status(400).json('Error: ' + err));    
+    AuthToken.findById(req.query.token )
+    .then((token) => {
+        const userId = req.body.userId; 
+        if (userId === token.userId) {
+            const remarkId = req.body.remarkId;
+            const content = req.body.content;
+            const newAnswer = new Answer({ userId, remarkId, content })
+            newAnswer.save()
+            .then(() => res.status(200).json('Answer added.'))
+            .catch(err => res.status(400).json('Error: ' + err));
+        } else {
+            res.status(400).json("Error: token doesn't match userId transmitted.")
+        }
+	})
+    .catch(err => {
+        res.status(401).json('Authentication Error: ' + err)
+    })
 });
 
 ////PUT REQUESTS
 
 //
 router.put('/up', (req,res,next) => {
-    Answer.findOneAndUpdate(
-    { 
-        _id: req.query.id
-    },
-    {
-        $inc : {ups : 1, pertinency: 1}
-    },
-    {useFindAndModify:false} //to avoid deprecation warning
-    )
-    .then(() => res.status(200).json('Answer up.'))
-    .catch(err => res.status(400).json('Error:' + err))
+    AuthToken.findById(req.query.token )
+    .then(
+        Answer.findOneAndUpdate(
+        { 
+            _id: req.query.id
+        },
+        {
+            $inc : {ups : 1, pertinency: 1}
+        },
+        {useFindAndModify:false} //to avoid deprecation warning
+        )
+        .then(() => res.status(200).json('Answer up.'))
+        .catch(err => res.status(400).json('Error:' + err))
+	)
+    .catch(err => {
+        res.status(401).json('Authentication Error: ' + err)
+    })
 });
 
 //
 router.put('/down', (req,res,next) => {
-    Answer.findOneAndUpdate(
-    { 
-        _id: req.query.id
-    },
-    {
-        $inc : {downs : 1, pertinency: -1}
-    },
-    {useFindAndModify:false} //to avoid deprecation warning
-    )
-    .then(() => res.status(200).json('Answer down.'))
-    .catch(err => res.status(400).json('Error:' + err))
+    AuthToken.findById(req.query.token )
+    .then(() => {
+        Answer.findOneAndUpdate(
+        { 
+            _id: req.query.id
+        },
+        {
+            $inc : {downs : 1, pertinency: -1}
+        },
+        {useFindAndModify:false} //to avoid deprecation warning
+        )
+        .then(() => res.status(200).json('Answer down.'))
+        .catch(err => res.status(400).json('Error:' + err))
+	})
+    .catch(err => {
+        res.status(401).json('Authentication Error: ' + err)
+    })
 });
 
 
@@ -112,11 +141,22 @@ router.put('/down', (req,res,next) => {
 
 //http://localhost:5000/answers/delete?id=5e591583dd55fc5880d34ca3
 router.delete('/delete', (req,res,next) => {
-    Answer.findOneAndDelete({ 
-        _id: req.query.id 
+    AuthToken.findById(req.query.token )
+    .then((token) => {
+        User.findById(token.userId)
+        .then((err,user) => {
+            if (user.admin) {
+                Answer.findOneAndDelete({ 
+                    _id: req.query.id 
+                })
+                .then((answer) => res.status(200).json("The following answer has been deleted by an administrator: " + answer))
+                .catch(err => res.status(400).json('Error:' + err))
+            } else {
+                res.status(403).json('Permission Error:' + err)
+            }
+        })
     })
-    .then((answer) => res.status(200).json("The answer has been deleted."))
-    .catch(err => res.status(400).json('Error:' + err))
+
 });
 
 module.exports = router;
